requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
flask>=3.0.0
python-telegram-bot>=20.0
schedule>=1.2.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
aiohttp>=3.9.0
pandas>=2.0.0
plotly>=5.18.0
